pyproject.toml
[project]
name = "consultant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.13.1",
    "httpx>=0.28.1",
    "langchain>=1.2.0",
    "langchain-openai>=1.1.4",
    "python-dotenv>=1.0.1",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
    "aiosmtplib>=3.0.1",
]
